requirements.txt
numpy
matplotlib
pillow
scikit-learn
pandas
seaborn
requests
ipykernel
tqdm
plotly
tensorflow==2.13.0
streamlit
opencv-python-headless
tensorflow-datasets
watchdog